refactor: rename filmdetails/playfilm packages, remove unused code
