+++ app/src/main/java/com/drs/auralife/data/local/dao/LibraryDao.kt
@@ -6,7 +6,6 @@ import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Transaction
 import com.drs.auralife.data.local.entity.LibraryEntity
-import com.drs.auralife.data.local.entity.LibraryFilmCrossRef
 
 @Dao
 interface LibraryDao {
