+++ app/src/main/java/com/drs/auralife/data/remote/firebase/AvatarDataSource.kt
@@ -6,7 +6,7 @@ import com.google.firebase.database.FirebaseDatabase
 import javax.inject.Inject
 
 class AvatarDataSource @Inject constructor(
-    private val database: FirebaseDatabase,
+    database: FirebaseDatabase,
 ) {
     private val userRef = database.getReference("users")
