+++ app/src/main/java/com/drs/auralife/data/remote/firebase/BannerDataSource.kt
@@ -5,7 +5,7 @@ import com.google.firebase.database.FirebaseDatabase
 import javax.inject.Inject
 
 class BannerDataSource @Inject constructor(
-    private val database: FirebaseDatabase,
+    database: FirebaseDatabase,
 ) {
     private val bannerRef = database.getReference("banners")
