+++ app/src/main/java/com/drs/auralife/presentation/common/MyAppGlideModule.kt
@@ -19,8 +19,7 @@ object MyAppGlideModule : AppGlideModule() {
             return false
         }
         if (context is Activity) {
-            val activity = context
-            if (activity.isDestroyed || activity.isFinishing) {
+            if (context.isDestroyed || context.isFinishing) {
                 return false
             }
         }
