+++ app/src/main/java/com/drs/auralife/presentation/explore_details/ExploreDetailsFragment.kt
@@ -7,17 +7,13 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
 import com.drs.auralife.presentation.common.launchAndRepeatWithViewLifecycle
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.GridLayoutManager
-import com.drs.auralife.R
 import com.drs.auralife.presentation.navigation.NavRoutes
 import com.drs.auralife.databinding.ActivityExploreDetailsBinding
 import com.drs.auralife.presentation.explore.adapter.ExploreFilmAdapter
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class ExploreDetailsFragment : Fragment() {
