+++ app/src/main/java/com/drs/auralife/presentation/library/LibraryFragment.kt
@@ -8,8 +8,6 @@ import android.widget.ImageButton
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
 import com.drs.auralife.presentation.common.launchAndRepeatWithViewLifecycle
 import androidx.navigation.fragment.findNavController
 import com.drs.auralife.R
@@ -18,7 +16,6 @@ import com.drs.auralife.databinding.FragmentLibraryBinding
 import com.drs.auralife.presentation.AppBarProvider
 import com.drs.auralife.presentation.library.adapter.LibraryAdapter
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class LibraryFragment : Fragment() {
