+++ app/src/main/res/layout/activity_film_details.xml
@@ -1,7 +1,7 @@
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.filmdetails.FilmDetailsActivity"
+    tools:context=".presentation.film_details.FilmDetailsActivity"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
