+++ app/src/main/res/layout/activity_play_film.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.playfilm.PlayFilmActivity">
+    tools:context=".presentation.play_film.PlayFilmActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
